Resolved conflict in maasauthpolicy_controller.go by merging OIDC TTL
and audiences configuration into the refactored gateway-level auth
policy structure from PRs #993, #981, #912.

Changes:
- Added audiences validation (oidc.ClientID) to JWT config
- Changed TTL from hardcoded 300 to dynamic oidc.TTL from Tenant CR
- Fixed type conversion to int64 for Kubernetes unstructured API compatibility

+++ maas-controller/pkg/controller/maas/maasauthpolicy_controller.go
@@ -644,7 +644,7 @@ func (r *MaaSAuthPolicyReconciler) buildGatewayAuthPolicySpec(modelAccessJSON st
 			"jwt": map[string]any{
 				"issuerUrl": oidc.IssuerURL,
 				"audiences": []any{oidc.ClientID},
-				"ttl":       oidc.TTL,
+				"ttl":       int64(oidc.TTL),
 			},
 			"when": []any{
 				map[string]any{
